requirements.txt
Django==2.2.7
beautifulsoup4==4.9.3
celery==4.4.2
certifi==2020.4.5.1
Django==2.2.7
djangorestframework==3.12.1
Jinja2==2.11.2
newsapi-python==0.2.6
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
redis==3.5.3
